+++ front/src/app/shared/models/shared.models.ts
@@ -1,0 +1,8 @@
+export interface TaskGroup  {
+  id: string;
+  title: string;
+}
+
+export interface TaskGroups {
+  lists: TaskGroup[];
+}
