--- front/src/app/shared/shared-interfaces.models.ts
@@ -1,12 +1,0 @@
-export class SharedInterfaces {
-
-}
-
-export interface TaskGroup  {
-  id: string;
-  title: string;
-}
-
-export interface TaskGroups {
-  lists: TaskGroup[];
-}
