+++ front/src/app/todo-list/todo-list.component.scss
@@ -1,46 +1,41 @@
-$active-task-group: rgb(235, 191, 69);
-
-.task-list::ng-deep {
+.task-list {
   max-width: 220px;
   margin-left: 20px;
   margin-top: 10px;
 
-  .task-list__group {
-    display: block;
+  &__link {
+    :link, :visited, :active {
+      color: black;
+    }
   }
 
-  .task-list__list {
+  &__list {
     width: 220px;
     height: 30px;
     border-radius: 25px;
     margin-top: 10px;
     text-align: center;
     display: block;
-  }
-  .task-list__list:hover {
-    background-color: rgb(216, 213, 213);
-  }
-  
-  .task-list__link:link, :visited, :active {
-    color: black;
-  }
-
-  .mdc-form-field {
-    vertical-align: -webkit-baseline-middle;
-  }
-
-  .mdc-label {
-    font-size: 16px;
-  }
-
-  .mdc-radio {
-    display: none;
+    &:hover {
+      background-color: rgb(216, 213, 213);
+    }
   }
 
   .mat-mdc-radio-checked {
-    background-color: $active-task-group;
+    background-color: rgb(235, 191, 69);
   }
-  .mat-mdc-radio-checked:hover {
-    background-color: $active-task-group;
+
+  ::ng-deep {
+    .mdc-form-field {
+      vertical-align: -webkit-baseline-middle;
+    }
+
+    .mdc-radio {
+      display: none;
+    }
+
+    .mdc-label {
+      font-size: 16px;
+    }
   }
 }
