+++ front/src/app/todo-list/todo-list.component.ts
@@ -1,8 +1,8 @@
 import { ChangeDetectionStrategy, ChangeDetectorRef, Component, OnInit, OnDestroy } from '@angular/core';
-import { ApiService } from '../shared/services/api.service';
+import { ApiService } from '../shared/api-service/api.service';
 import { FormControl, FormGroup } from '@angular/forms';
 import { ActivatedRoute } from '@angular/router';
-import { TaskGroups } from '../shared/shared-interfaces.models';
+import { TaskGroups } from '../shared/models/shared.models';
 import { Subscription } from 'rxjs';
 
 @Component({
@@ -12,20 +12,20 @@ import { Subscription } from 'rxjs';
   changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class TodoListComponent implements OnInit, OnDestroy { 
-  allTaskLists: TaskGroups| undefined;
+  allTaskLists: TaskGroups | undefined;
 
   radioBtnGroup: FormGroup = new FormGroup ({
     activeTaskList: new FormControl()
   })
 
-  subscription: Subscription | undefined;
+  subscriptionToTaskList: Subscription | undefined;
 
   constructor(private apiService: ApiService, private changeDetectorRef: ChangeDetectorRef, private route: ActivatedRoute) {}
 
   ngOnInit(): void {
     this.loadTaskLists();
 
-    this.subscription = this.route.queryParams
+    this.route.queryParams
       .subscribe(params => {
         if (params['list']) {
           this.radioBtnGroup.controls['activeTaskList'].setValue(params['list'])
@@ -34,11 +34,11 @@ export class TodoListComponent implements OnInit, OnDestroy {
   }
 
   ngOnDestroy(): void {
-    this.subscription?.unsubscribe();
+    this.subscriptionToTaskList?.unsubscribe();
   }
 
   private loadTaskLists() {
-    this.apiService.getLists()
+    this.subscriptionToTaskList = this.apiService.getLists()
       .subscribe(data => {
         this.allTaskLists = data;
         this.changeDetectorRef.markForCheck();
